backend/pyproject.toml
[project]
name = "pda-nbr5419-backend"
version = "0.1.0"
description = "Sistema PDA - Analise de risco e laudos tecnicos conforme ABNT NBR 5419:2026"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.2.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.0,<5.0",
    "python-multipart>=0.0.12",
    "weasyprint>=62.0",
    "jinja2>=3.1.4",
    "pillow>=10.4.0",
    "python-docx>=1.1.0",
    "psycopg2-binary",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]